requirements.txt
torchvision
numpy
matplotlib
omegaconf
opencv_python_headless
pandas
pytorch_lightning
albumentations
hydra_core
comet_ml
tqdm
torch
pycocotools
Pillow
scikit_learn
tensorboard